+++ example/shieldwall/endcap.py
@@ -15,4 +15,4 @@ cap_bp.make()
 cap_ex = cap_bp.build()
 
 #show_object(cap_ex)
-cq.exporters.export(cap_ex, 'stl/endcap.stl')
+cq.exporters.export(cap_ex, 'stl/shieldwall_endcap.stl')