+++ example/shieldwall/shape.py
@@ -7,4 +7,4 @@ shape_bp.make()
 shape_ex = shape_bp.build().extrude(2)
 #show_object(shape_ex)
 
-cq.exporters.export(shape_ex, 'stl/shieldwallshape.stl')
+cq.exporters.export(shape_ex, 'stl/shieldwall_shape.stl')