+++ example/shieldwall/straight_basic.py
@@ -6,4 +6,4 @@ wall_bp.make()
 wall_ex = wall_bp.build()
 
 #show_object(wall_ex)
-cq.exporters.export(wall_ex, 'stl/straightbasic.stl')
+cq.exporters.export(wall_ex, 'stl/shieldwall_straightbasic.stl')