+++ example/shieldwall/straight_hex.py
@@ -19,6 +19,6 @@ straight_ex = straight_bp.build()
 cross_section = cq.Workplane('XY').box(10,straight_bp.width,50)
 
 #show_object(straight_ex)
-cq.exporters.export(straight_ex, 'stl/straight_hex.stl')
+cq.exporters.export(straight_ex, 'stl/shieldwall_straight_hex.stl')